requirements.txt
telnyx
fastapi
uvicorn[standard]
python-dotenv